+++ src/hooks/use-tickets-mutations.ts
@@ -1,0 +1,21 @@
+import { ticketService } from "@/services/tickets.service";
+import { TicketResaleRequest } from "@/types/ticket";
+import { useMutation} from "@tanstack/react-query";
+import { toast } from "sonner";
+
+export function useTicketResale() {
+
+    return useMutation({
+        mutationKey: ['ticket-resale'],
+        mutationFn: async (data: TicketResaleRequest[]) => {
+            await ticketService.resellTickets(data);
+        },
+        onSuccess: () => {
+            toast.success('Ticket resale successful.');
+        },
+        onError: () => {
+            toast.error('Failed to resell tickets. Please try again.');
+        },
+    })
+
+}